SecureTalk + steganografija/requirements.txt
Pillow>=9.0.0
cryptography>=41.0.0
websockets>=11.0.0
pynacl>=1.5.0
numpy>=1.24.0
sounddevice>=0.4.6
fastapi>=0.95.0
uvicorn>=0.22.0
